Add concurrent download progress

+++ Library/Homebrew/download_queue.rb
@@ -27,6 +27,8 @@ module Homebrew
       @force = force
       @pour = pour
       @pool = T.let(Concurrent::FixedThreadPool.new(concurrency), Concurrent::FixedThreadPool)
+      @tty = T.let($stdout.tty?, T::Boolean)
+      @spinner = T.let(nil, T.nilable(Spinner))
     end
 
     sig {
@@ -65,39 +67,20 @@ module Homebrew
           raise e unless bottle_manifest_error?(downloadable, e)
         end
       else
-        spinner = Spinner.new
         remaining_downloads = downloads.dup.to_a
         previous_pending_line_count = 0
-        tty = $stdout.tty?
 
         begin
           stdout_print_and_flush_if_tty Tty.hide_cursor
 
           output_message = lambda do |downloadable, future, last|
-            status = case future.state
-            when :fulfilled
-              if tty
-                "#{Tty.green}✔︎#{Tty.reset}"
-              else
-                "✔︎"
-              end
-            when :rejected
-              if tty
-                "#{Tty.red}✘#{Tty.reset}"
-              else
-                "✘"
-              end
-            when :pending, :processing
-              "#{Tty.blue}#{spinner}#{Tty.reset}" if tty
-            else
-              raise future.state.to_s
-            end
-
+            status = status_from_future(future)
             exception = future.reason if future.rejected?
             next 1 if bottle_manifest_error?(downloadable, exception)
 
             message = "#{downloadable.download_queue_type} #{downloadable.download_queue_name}"
             if tty
+              message = message_with_progress(downloadable, future, message)
               stdout_print_and_flush "#{status} #{message}#{"\n" unless last}"
             elsif status
               $stderr.puts "#{status} #{message}"
@@ -111,9 +94,8 @@ module Homebrew
                 Homebrew.failed = true if downloadable.is_a?(Resource::Patch)
                 next 2
               elsif exception.is_a?(CannotInstallFormulaError)
-                if (cached_download = downloadable.cached_download)&.exist?
-                  cached_download.unlink
-                end
+                cached_download = downloadable.cached_download
+                cached_download.unlink if cached_download&.exist?
                 raise exception
               else
                 message = future.reason.to_s
@@ -237,9 +219,72 @@ module Homebrew
     sig { returns(T::Boolean) }
     attr_reader :pour
 
+    sig { returns(T::Boolean) }
+    attr_reader :tty
+
     sig { returns(T::Hash[Downloadable, Concurrent::Promises::Future]) }
     def downloads
       @downloads ||= T.let({}, T.nilable(T::Hash[Downloadable, Concurrent::Promises::Future]))
+    end
+
+    sig { params(future: Concurrent::Promises::Future).returns(T.nilable(String)) }
+    def status_from_future(future)
+      case future.state
+      when :fulfilled
+        if tty
+          "#{Tty.green}✔︎#{Tty.reset}"
+        else
+          "✔︎"
+        end
+      when :rejected
+        if tty
+          "#{Tty.red}✘#{Tty.reset}"
+        else
+          "✘"
+        end
+      when :pending, :processing
+        "#{Tty.blue}#{spinner}#{Tty.reset}" if tty
+      else
+        raise future.state.to_s
+      end
+    end
+
+    sig { returns(Spinner) }
+    def spinner
+      @spinner ||= Spinner.new
+    end
+
+    sig { params(downloadable: Downloadable, future: Concurrent::Promises::Future, message: String).returns(String) }
+    def message_with_progress(downloadable, future, message)
+      tty_width = Tty.width
+      return message unless tty_width.positive?
+
+      fetched_size = downloadable.fetched_size
+      return message if fetched_size.blank?
+
+      size_length = 5
+      unit_length = 2
+      size_formatting_string = "%<size>#{size_length}.1f%<unit>#{unit_length}s"
+      size, unit = disk_usage_readable_size_unit(fetched_size)
+      formatted_fetched_size = format(size_formatting_string, size:, unit:)
+
+      formatted_total_size = if future.fulfilled?
+        formatted_fetched_size
+      elsif (total_size = downloadable.total_size)
+        size, unit = disk_usage_readable_size_unit(total_size)
+        format(size_formatting_string, size:, unit:)
+      else
+        # fill in the missing spaces for the size if we don't have it yet.
+        "-" * (size_length + unit_length)
+      end
+
+      max_phase_length = 11
+      phase = format("%-<phase>#{max_phase_length}s", phase: downloadable.phase.to_s.capitalize)
+      progress = "#{formatted_fetched_size}/#{formatted_total_size}"
+      additional_padding_length = max_phase_length + size_length + unit_length - 1
+      message_length = tty_width - progress.length - additional_padding_length
+
+      "#{message[0, message_length].to_s.ljust(message_length)} [#{phase} #{progress}]"
     end
 
     class Spinner
